+++ app/Http/Controllers/Master/BoardController.php
@@ -54,12 +54,26 @@ class BoardController extends Controller
                 return $created_at;
             })
               ->addColumn('action', function ($row) {
-                $edit_btn = '<a href="javascript:void(0);" onclick="getBoardModal(' . $row->id . ')"  class="btn btn-icon btn-active-primary btn-light-primary mx-1 w-30px h-30px" > 
-                <i class="fa fa-edit"></i>
-            </a>';
+                $route_name = request()->route()->getName(); 
+                if( access()->buttonAccess($route_name,'add_edit') )
+                {
+                    $edit_btn = '<a href="javascript:void(0);" onclick="getBoardModal(' . $row->id . ')"  class="btn btn-icon btn-active-primary btn-light-primary mx-1 w-30px h-30px" > 
+                    <i class="fa fa-edit"></i>
+                    </a>';
+                }
+                else
+                {
+                    $edit_btn = '';
+                }
+                if( access()->buttonAccess($route_name,'delete') )
+                {
                     $del_btn = '<a href="javascript:void(0);" onclick="deleteBoard(' . $row->id . ')" class="btn btn-icon btn-active-danger btn-light-danger mx-1 w-30px h-30px" > 
-                <i class="fa fa-trash"></i></a>';
-
+                    <i class="fa fa-trash"></i></a>';
+                }
+                else
+                {
+                    $del_btn = '';
+                }
                     return $edit_btn . $del_btn;
                 })
                 ->rawColumns(['action', 'status']);
